refactor(email): pull selectEmailServices out to its own module
