+++ packages/fxa-auth-server/test/local/senders/email.js
@@ -7,14 +7,10 @@
 const ROOT_DIR = '../../..';
 
 const { assert } = require('chai');
-const cp = require('child_process');
 const mocks = require('../../mocks');
 const P = require('bluebird');
-const path = require('path');
-const proxyquire = require('proxyquire').noPreserveCache();
+const proxyquire = require('proxyquire');
 const sinon = require('sinon');
-
-cp.execAsync = P.promisify(cp.exec);
 
 const config = require(`${ROOT_DIR}/config`).getProperties();
 if (!config.smtp.prependVerificationSubdomain.enabled) {
@@ -806,20 +802,16 @@ const TRAILHEAD_TESTS = new Map([
 ]);
 
 describe('lib/senders/email:', () => {
-  let mockLog, redis, mailer, localize, selectEmailServices, sendMail;
+  let mockLog, mailer, localize, selectEmailServices, sendMail;
 
   before(async () => {
     mockLog = mocks.mockLog();
-    redis = {
-      get: sinon.spy(() => P.resolve()),
-    };
     mailer = await setup(mockLog, config, {
       './oauth_client_info': () => ({
         async fetch() {
           return { name: 'Mock Relier' };
         },
       }),
-      '../redis': () => redis,
     });
     // These tests do a lot of ad hoc mocking. Rather than try and clean up
     // after each case, give them carte blanche to do what they want then
@@ -840,7 +832,6 @@ describe('lib/senders/email:', () => {
         fn.resetHistory();
       }
     });
-    redis.get.resetHistory();
     if (mailer.localize !== localize) {
       mailer.localize = localize;
     }
@@ -1141,8 +1132,6 @@ describe('lib/senders/email:', () => {
         assert.equal(headers['X-SES-CONFIGURATION-SET'], 'wibble');
 
         assert.equal(typeof args[1], 'function');
-
-        assert.equal(redis.get.callCount, 1);
       });
     });
 
@@ -1165,8 +1154,6 @@ describe('lib/senders/email:', () => {
         assert.equal(args[0].headers['X-Uid'], 'foo');
         assert.equal(args[0].provider, undefined);
         assert.equal(typeof mailer.mailer.sendMail.args[0][1], 'function');
-
-        assert.equal(redis.get.callCount, 1);
       });
     });
 
@@ -1309,793 +1296,6 @@ describe('lib/senders/email:', () => {
       });
     });
   });
-
-  describe('single email address:', () => {
-    const emailAddress = 'foo@example.com';
-
-    describe('redis.get returns sendgrid percentage-only match:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(JSON.stringify({ sendgrid: { percentage: 11 } }))
-        );
-        sinon.stub(Math, 'random').callsFake(() => 0.109);
-      });
-
-      afterEach(() => Math.random.restore());
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.emailService,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-email-service',
-                emailSender: 'sendgrid',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns sendgrid percentage-only mismatch:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(JSON.stringify({ sendgrid: { percentage: 11 } }))
-        );
-        sinon.stub(Math, 'random').callsFake(() => 0.11);
-      });
-
-      afterEach(() => Math.random.restore());
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.mailer,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns sendgrid regex-only match:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(
-            JSON.stringify({ sendgrid: { regex: '^foo@example.com$' } })
-          )
-        );
-      });
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.emailService,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-email-service',
-                emailSender: 'sendgrid',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns sendgrid regex-only mismatch:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(JSON.stringify({ sendgrid: { regex: '^fo@example.com$' } }))
-        );
-      });
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.mailer,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns sendgrid combined match:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(
-            JSON.stringify({
-              sendgrid: {
-                percentage: 1,
-                regex: '^foo@example.com$',
-              },
-            })
-          )
-        );
-        sinon.stub(Math, 'random').callsFake(() => 0.009);
-      });
-
-      afterEach(() => Math.random.restore());
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.emailService,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-email-service',
-                emailSender: 'sendgrid',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns sendgrid combined mismatch (percentage):', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(
-            JSON.stringify({
-              sendgrid: {
-                percentage: 1,
-                regex: '^foo@example.com$',
-              },
-            })
-          )
-        );
-        sinon.stub(Math, 'random').callsFake(() => 0.01);
-      });
-
-      afterEach(() => Math.random.restore());
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.mailer,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns sendgrid combined mismatch (regex):', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(
-            JSON.stringify({
-              sendgrid: {
-                percentage: 1,
-                regex: '^ffoo@example.com$',
-              },
-            })
-          )
-        );
-        sinon.stub(Math, 'random').callsFake(() => 0);
-      });
-
-      afterEach(() => Math.random.restore());
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.mailer,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns socketlabs percentage-only match:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(JSON.stringify({ socketlabs: { percentage: 42 } }))
-        );
-        sinon.stub(Math, 'random').callsFake(() => 0.419);
-      });
-
-      afterEach(() => Math.random.restore());
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.emailService,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-email-service',
-                emailSender: 'socketlabs',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns socketlabs percentage-only mismatch:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(JSON.stringify({ socketlabs: { percentage: 42 } }))
-        );
-        sinon.stub(Math, 'random').callsFake(() => 0.42);
-      });
-
-      afterEach(() => Math.random.restore());
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.mailer,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns socketlabs regex-only match:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(
-            JSON.stringify({ socketlabs: { regex: '^foo@example.com$' } })
-          )
-        );
-      });
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.emailService,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-email-service',
-                emailSender: 'socketlabs',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns ses percentage-only match:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(JSON.stringify({ ses: { percentage: 100 } }))
-        );
-        sinon.stub(Math, 'random').callsFake(() => 0.999);
-      });
-
-      afterEach(() => Math.random.restore());
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.emailService,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-email-service',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns ses percentage-only mismatch:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(JSON.stringify({ ses: { percentage: 99 } }))
-        );
-        sinon.stub(Math, 'random').callsFake(() => 0.999);
-      });
-
-      afterEach(() => Math.random.restore());
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.mailer,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns ses regex-only match:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(JSON.stringify({ ses: { regex: '^foo@example.com$' } }))
-        );
-      });
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.emailService,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-email-service',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns sendgrid and ses matches:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(
-            JSON.stringify({
-              sendgrid: { percentage: 10 },
-              ses: { regex: '^foo@example.com$' },
-            })
-          )
-        );
-        sinon.stub(Math, 'random').callsFake(() => 0.09);
-      });
-
-      afterEach(() => Math.random.restore());
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.emailService,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-email-service',
-                emailSender: 'sendgrid',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns sendgrid match and ses mismatch:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(
-            JSON.stringify({
-              sendgrid: { percentage: 10 },
-              ses: { regex: '^ffoo@example.com$' },
-            })
-          )
-        );
-        sinon.stub(Math, 'random').callsFake(() => 0.09);
-      });
-
-      afterEach(() => Math.random.restore());
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.emailService,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-email-service',
-                emailSender: 'sendgrid',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns sendgrid mismatch and ses match:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(
-            JSON.stringify({
-              sendgrid: { percentage: 10 },
-              ses: { regex: '^foo@example.com$' },
-            })
-          )
-        );
-        sinon.stub(Math, 'random').callsFake(() => 0.1);
-      });
-
-      afterEach(() => Math.random.restore());
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.emailService,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-email-service',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns sendgrid and ses mismatches:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(
-            JSON.stringify({
-              sendgrid: { percentage: 10 },
-              ses: { regex: '^ffoo@example.com$' },
-            })
-          )
-        );
-        sinon.stub(Math, 'random').callsFake(() => 0.1);
-      });
-
-      afterEach(() => Math.random.restore());
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.mailer,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns undefined:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() => P.resolve());
-      });
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.mailer,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns unsafe regex:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(
-            JSON.stringify({ sendgrid: { regex: '^(.+)+@example.com$' } })
-          )
-        );
-      });
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.mailer,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns quote-terminating regex:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(JSON.stringify({ sendgrid: { regex: '"@example.com$' } }))
-        );
-      });
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.mailer,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('email address contains quote-terminator:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(JSON.stringify({ sendgrid: { regex: '@example.com$' } }))
-        );
-      });
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: '"@example.com' })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.mailer,
-                emailAddresses: ['"@example.com'],
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get fails:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() => P.reject({ message: 'wibble' }));
-      });
-
-      it('selectEmailServices returns fallback data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result => {
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.mailer,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ]);
-            assert.equal(mockLog.error.callCount, 1);
-            const args = mockLog.error.args[0];
-            assert.equal(args.length, 2);
-            assert.equal(args[0], 'emailConfig.read.error');
-            assert.deepEqual(args[1], {
-              err: 'wibble',
-            });
-          });
-      });
-    });
-
-    describe('redis.get returns invalid JSON:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() => P.resolve('wibble'));
-      });
-
-      it('selectEmailServices returns fallback data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result => {
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.mailer,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ]);
-            assert.equal(mockLog.error.callCount, 1);
-            assert.equal(mockLog.error.args[0][0], 'emailConfig.parse.error');
-          });
-      });
-    });
-  });
-
-  describe('single email address matching local static email service config:', () => {
-    const emailAddress = 'emailservice.1@restmail.net';
-
-    describe('redis.get returns sendgrid match:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(JSON.stringify({ sendgrid: { regex: 'restmail' } }))
-        );
-      });
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.emailService,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-email-service',
-                emailSender: 'sendgrid',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns sendgrid mismatch:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(JSON.stringify({ sendgrid: { regex: 'rustmail' } }))
-        );
-      });
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({ email: emailAddress })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.emailService,
-                emailAddresses: [emailAddress],
-                emailService: 'fxa-email-service',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-  });
-
-  describe('multiple email addresses:', () => {
-    const emailAddresses = ['a@example.com', 'b@example.com', 'c@example.com'];
-
-    describe('redis.get returns sendgrid and ses matches and a mismatch:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(
-            JSON.stringify({
-              sendgrid: { regex: '^a' },
-              ses: { regex: '^b' },
-            })
-          )
-        );
-      });
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({
-            email: emailAddresses[0],
-            ccEmails: emailAddresses.slice(1),
-          })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.emailService,
-                emailAddresses: emailAddresses.slice(0, 1),
-                emailService: 'fxa-email-service',
-                emailSender: 'sendgrid',
-              },
-              {
-                mailer: mailer.emailService,
-                emailAddresses: emailAddresses.slice(1, 2),
-                emailService: 'fxa-email-service',
-                emailSender: 'ses',
-              },
-              {
-                mailer: mailer.mailer,
-                emailAddresses: emailAddresses.slice(2),
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns a sendgrid match and two ses matches:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(
-            JSON.stringify({
-              sendgrid: { regex: '^a' },
-              ses: { regex: '^b|c' },
-            })
-          )
-        );
-      });
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({
-            email: emailAddresses[0],
-            ccEmails: emailAddresses.slice(1),
-          })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.emailService,
-                emailAddresses: emailAddresses.slice(0, 1),
-                emailService: 'fxa-email-service',
-                emailSender: 'sendgrid',
-              },
-              {
-                mailer: mailer.emailService,
-                emailAddresses: emailAddresses.slice(1),
-                emailService: 'fxa-email-service',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-
-    describe('redis.get returns three mismatches:', () => {
-      beforeEach(() => {
-        redis.get = sinon.spy(() =>
-          P.resolve(
-            JSON.stringify({
-              sendgrid: { regex: 'wibble' },
-              ses: { regex: 'blee' },
-            })
-          )
-        );
-      });
-
-      it('selectEmailServices returns the correct data', () => {
-        return mailer
-          .selectEmailServices({
-            email: emailAddresses[0],
-            ccEmails: emailAddresses.slice(1),
-          })
-          .then(result =>
-            assert.deepEqual(result, [
-              {
-                mailer: mailer.mailer,
-                emailAddresses: emailAddresses,
-                emailService: 'fxa-auth-server',
-                emailSender: 'ses',
-              },
-            ])
-          );
-      });
-    });
-  });
 });
 
 describe('mailer constructor:', () => {
@@ -2148,170 +1348,15 @@ describe('mailer constructor:', () => {
   });
 });
 
-describe('call selectEmailServices with mocked sandbox:', () => {
-  const emailAddress = 'foo@example.com';
-  let mockLog, redis, Sandbox, sandbox, mailer, promise, result, failed;
-
-  before(async () => {
-    mockLog = mocks.mockLog();
-    redis = {
-      get: sinon.spy(() =>
-        P.resolve(JSON.stringify({ sendgrid: { regex: '^foo@example.com$' } }))
-      ),
-    };
-    // eslint-disable-next-line prefer-arrow-callback
-    Sandbox = sinon.spy(function() {
-      return sandbox;
-    });
-    mailer = await setup(mockLog, config, {
-      '../redis': () => redis,
-      sandbox: Sandbox,
-    });
-  });
-
-  after(() => mailer.stop());
-
-  beforeEach(done => {
-    sandbox = {
-      run: sinon.spy(),
-    };
-    promise = mailer
-      .selectEmailServices({ email: emailAddress })
-      .then(r => (result = r))
-      .catch(() => (failed = true));
-    setImmediate(done);
-  });
-
-  it('called the sandbox correctly', () => {
-    assert.equal(Sandbox.callCount, 1);
-
-    let args = Sandbox.args[0];
-    assert.equal(args.length, 1);
-    assert.deepEqual(args[0], {
-      timeout: 100,
-    });
-
-    assert.equal(sandbox.run.callCount, 1);
-
-    args = sandbox.run.args[0];
-    assert.equal(args.length, 2);
-    assert.equal(
-      args[0],
-      'new RegExp("^foo@example.com$").test("foo@example.com")'
-    );
-    assert.equal(typeof args[1], 'function');
-  });
-
-  describe('call sandbox result handler with match:', () => {
-    beforeEach(() => {
-      sandbox.run.args[0][1]({ result: 'true' });
-      return promise;
-    });
-
-    it('resolved', () => {
-      assert.deepEqual(result, [
-        {
-          emailAddresses: ['foo@example.com'],
-          mailer: mailer.emailService,
-          emailService: 'fxa-email-service',
-          emailSender: 'sendgrid',
-        },
-      ]);
-    });
-
-    it('did not fail', () => {
-      assert.equal(failed, undefined);
-    });
-  });
-
-  describe('call sandbox result handler with timeout:', () => {
-    beforeEach(() => {
-      sandbox.run.args[0][1]({ result: 'TimeoutError' });
-      return promise;
-    });
-
-    it('resolved', () => {
-      assert.deepEqual(result, [
-        {
-          emailAddresses: ['foo@example.com'],
-          mailer: mailer.mailer,
-          emailService: 'fxa-auth-server',
-          emailSender: 'ses',
-        },
-      ]);
-    });
-
-    it('did not fail', () => {
-      assert.equal(failed, undefined);
-    });
-  });
-});
-
-describe('call selectEmailServices with mocked safe-regex, regex-only match and redos regex:', () => {
-  const emailAddress = 'foo@example.com';
-  let mockLog, redis, safeRegex, mailer;
-
-  before(async () => {
-    mockLog = mocks.mockLog();
-    redis = {
-      get: sinon.spy(() =>
-        P.resolve(
-          JSON.stringify({
-            sendgrid: { regex: '^((((.*)*)*)*)*@example.com$' },
-          })
-        )
-      ),
-    };
-    // eslint-disable-next-line prefer-arrow-callback
-    safeRegex = sinon.spy(function() {
-      return true;
-    });
-    mailer = await setup(mockLog, config, {
-      '../redis': () => redis,
-      'safe-regex': safeRegex,
-    });
-  });
-
-  after(() => mailer.stop());
-
-  it('email address was treated as mismatch', () => {
-    return mailer.selectEmailServices({ email: emailAddress }).then(result => {
-      assert.deepEqual(result, [
-        {
-          mailer: mailer.mailer,
-          emailAddresses: [emailAddress],
-          emailService: 'fxa-auth-server',
-          emailSender: 'ses',
-        },
-      ]);
-
-      assert.equal(safeRegex.callCount, 1);
-      const args = safeRegex.args[0];
-      assert.equal(args.length, 1);
-      assert.equal(args[0], '^((((.*)*)*)*)*@example.com$');
-    });
-  });
-});
-
 describe('email translations', () => {
-  let mockLog, redis, mailer;
+  let mockLog, mailer;
   const message = {
     email: 'a@b.com',
   };
 
   async function setupMailerWithTranslations(locale) {
     mockLog = mocks.mockLog();
-    redis = {
-      get: sinon.spy(() => P.resolve()),
-    };
-    mailer = await setup(
-      mockLog,
-      config,
-      {
-        '../redis': () => redis,
-      },
-      locale
-    );
+    mailer = await setup(mockLog, config, {}, locale);
   }
 
   afterEach(() => mailer.stop());
@@ -2350,53 +1395,6 @@ describe('email translations', () => {
     });
   });
 });
-
-if (config.redis.email.enabled) {
-  const emailAddress = 'foo@example.com';
-
-  ['sendgrid', 'ses', 'socketlabs'].reduce(async (promise, service) => {
-    await promise;
-
-    return new P((resolve, reject) => {
-      describe(`call selectEmailServices with real redis containing ${service} config:`, () => {
-        let mailer, result;
-
-        before(async () => {
-          const mockLog = mocks.mockLog();
-          mailer = await setup(mockLog, config, {});
-          await redisWrite({
-            [service]: {
-              regex: '^foo@example.com$',
-              percentage: 100,
-            },
-          });
-          result = await mailer.selectEmailServices({ email: emailAddress });
-        });
-
-        after(async () => {
-          try {
-            await redisRevert();
-            await mailer.stop();
-            resolve();
-          } catch (err) {
-            reject(err);
-          }
-        });
-
-        it('returned the correct result', () => {
-          assert.deepEqual(result, [
-            {
-              emailAddresses: [emailAddress],
-              emailService: 'fxa-email-service',
-              emailSender: service,
-              mailer: mailer.emailService,
-            },
-          ]);
-        });
-      });
-    });
-  }, P.resolve());
-}
 
 function sesMessageTagsHeaderValue(templateName, serviceName) {
   return `messageType=fxa-${templateName}, app=fxa, service=${serviceName ||
@@ -2490,20 +1488,5 @@ function applyAssertions(type, target, property, assertions) {
 
   assertions.forEach(({ test, expected }) => {
     assert[test](target, expected, `${type}: ${property}`);
-  });
-}
-
-function redisWrite(config) {
-  return cp.execAsync(
-    `echo '${JSON.stringify(config)}' | node scripts/email-config write`,
-    {
-      cwd: path.resolve(__dirname, '../../..'),
-    }
-  );
-}
-
-function redisRevert() {
-  return cp.execAsync('node scripts/email-config revert', {
-    cwd: path.resolve(__dirname, '../../..'),
   });
 }
